+++ packages/expo-camera/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 💡 Others
 
 - Updated `@expo/config-plugins` from `4.0.2` to `4.0.14` ([#15621](https://github.com/expo/expo/pull/15621) by [@EvanBacon](https://github.com/EvanBacon))
+- Replace `CapturedPicture` type with `CameraCapturedPicture` in events callback to avoid duplicated types.
 
 ## 12.1.0 — 2021-12-03